pyproject.toml
[tool.poetry]
name = "flask-mega-tutorial"
version = "0.1.0"
description = ""
authors = ["Joe Matusik <jpmatusik@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
Flask-WTF = "~0.14.3"
Flask-SQLAlchemy = "^2.4.4"
Flask-Login = "~0.5.0"
Flask-Bcrypt = "~0.7.1"
psycopg2 = "^2.8.6"
email-validator = "^1.1.2"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
Flask-Migrate = "^2.5.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
src_paths = ["flask_mega_tutorial"]
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
